fix: WS-3 claims unconvert; and unused should not be enabled at all (#313)
**Claiming `unconvert`.** 32 findings, not the 23 recorded -- and eight of the
extra were created hours earlier by staticcheck's SA4016 fix in the same files.
Replacing `uint64(written)<<32 | 0` with `packSimpleResult(0, uint32(written))`
introduced a redundant conversion, because `written` was already uint32. Two
linters disagreeing about the same lines until both are run is an argument for
running them together rather than one per session, and worth knowing before the
next person takes one off the list.

Applied with `unconvert -apply ./...` rather than by hand. One judgement call:
its single edit to a _test.go file is reverted, because this config excludes
test files and the change should be exactly what the linter enforces.

Nothing here is behavioural. The one that deserved a second look was in cgo --
`C.uint64_t(instance.store_id)` where the field is already C.uint64_t. Removing
it is if anything safer: with the conversion, an upstream type change still
compiles; without it, the build breaks and tells you.

**unused: measured at 30 and deliberately NOT enabled.**

It is the same U1000 analysis that scripts/check-test-only-code.sh already runs,
against its own 59-entry baseline. Enabling it here would give one class of
finding two mechanisms with two baselines that can disagree -- which is the
shape that let the two routing tables in 2.72 drift apart, and it took a
session to untangle. If dead code should be enforced by golangci-lint, the move
is to retire the script, not to run both. Recorded in .golangci.yml so the next
person does not spend the measurement again.

That leaves errcheck (280), gosec (283) and gocyclo (35) on the list. The first
two are the ones PARALLEL-WORKSTREAMS.md says matter and should not be deferred
forever on size; gocyclo is a refactoring project rather than a sweep.

Verified: `go build ./...` clean; `golangci-lint run` clean repo-wide with four
linters now enabled; `go test` green for engine (three dialects) and
cleat-worker.

+++ cmd/cleat-worker/setup.go
@@ -1909,7 +1909,7 @@ func (w *Worker) concurrencyKeyReaperLoop() {
 			}
 			if reaped > 0 {
 				w.logger.InfoContext(w.ctx, "Concurrency key reaper: removed expired keys", "worker_id", w.id, "count", reaped)
-				w.Metrics.SetBackgroundLoopItemsProcessed(w.ctx, "concurrency_key_reaper", int64(reaped))
+				w.Metrics.SetBackgroundLoopItemsProcessed(w.ctx, "concurrency_key_reaper", reaped)
 			}
 			w.Metrics.RecordBackgroundLoop(w.ctx, "concurrency_key_reaper", "ok")
 			w.Metrics.SetBackgroundLoopDuration(w.ctx, "concurrency_key_reaper", time.Since(ckStart).Seconds())
@@ -2097,7 +2097,7 @@ func (w *Worker) memoryCleanupLoop(maxSamples int) {
 			if err == nil {
 				w.Metrics.RecordBackgroundLoop(w.ctx, "memory_cleanup", "ok")
 				if deleted > 0 {
-					w.Metrics.SetBackgroundLoopItemsProcessed(w.ctx, "memory_cleanup", int64(deleted))
+					w.Metrics.SetBackgroundLoopItemsProcessed(w.ctx, "memory_cleanup", deleted)
 				}
 			}
 			w.Metrics.SetBackgroundLoopDuration(w.ctx, "memory_cleanup", time.Since(mcStart).Seconds())
