+++ engine/signaller.go
@@ -20,7 +20,7 @@ func (s *execSession) DurableAwaitSignals(ctx context.Context, m api.Module, sig
 				}
 				written, _ := s.writeResult(ctx, m, sigNamePtr, rec.SignalName, sigNameMaxLen)
 				_, _ = s.writeResult(ctx, m, payloadPtr, rec.SignalPayload, payloadMaxLen)
-				return packAwaitSignalsResult(uint32(written), uint32(len(rec.SignalPayload)), false, 0)
+				return packAwaitSignalsResult(written, uint32(len(rec.SignalPayload)), false, 0)
 			}
 			if rec.EventType == EventTypeAwaitSignals {
 				if !s.advanceReplayStep(ctx, &rec) {
@@ -35,7 +35,7 @@ func (s *execSession) DurableAwaitSignals(ctx context.Context, m api.Module, sig
 						}
 						written, _ := s.writeResult(ctx, m, sigNamePtr, nextRec.SignalName, sigNameMaxLen)
 						_, _ = s.writeResult(ctx, m, payloadPtr, nextRec.SignalPayload, payloadMaxLen)
-						return packAwaitSignalsResult(uint32(written), uint32(len(nextRec.SignalPayload)), false, 0)
+						return packAwaitSignalsResult(written, uint32(len(nextRec.SignalPayload)), false, 0)
 					}
 				}
 				// No signal_received in history. The signal may have
@@ -61,7 +61,7 @@ func (s *execSession) DurableAwaitSignals(ctx context.Context, m api.Module, sig
 							s.recordEvent(sigRec)
 							written, _ := s.writeResult(ctx, m, sigNamePtr, name, sigNameMaxLen)
 							_, _ = s.writeResult(ctx, m, payloadPtr, payload, payloadMaxLen)
-							return packAwaitSignalsResult(uint32(written), uint32(len(payload)), false, 0)
+							return packAwaitSignalsResult(written, uint32(len(payload)), false, 0)
 						}
 					}
 				}
@@ -90,7 +90,7 @@ func (s *execSession) DurableAwaitSignals(ctx context.Context, m api.Module, sig
 
 				written, _ := s.writeResult(ctx, m, sigNamePtr, name, sigNameMaxLen)
 				_, _ = s.writeResult(ctx, m, payloadPtr, payload, payloadMaxLen)
-				return packAwaitSignalsResult(uint32(written), uint32(len(payload)), false, 0)
+				return packAwaitSignalsResult(written, uint32(len(payload)), false, 0)
 			}
 		}
 	}
